Move unmigrated models to `models/legacy`
