+++ keras_cv/models/__init__.py
@@ -100,38 +100,6 @@ from keras_cv.models.backbones.resnet_v2.resnet_v2_backbone import (
     ResNetV2Backbone,
 )
 from keras_cv.models.classification.image_classifier import ImageClassifier
-from keras_cv.models.convmixer import ConvMixer_512_16
-from keras_cv.models.convmixer import ConvMixer_768_32
-from keras_cv.models.convmixer import ConvMixer_1024_16
-from keras_cv.models.convmixer import ConvMixer_1536_20
-from keras_cv.models.convmixer import ConvMixer_1536_24
-from keras_cv.models.convnext import ConvNeXtBase
-from keras_cv.models.convnext import ConvNeXtLarge
-from keras_cv.models.convnext import ConvNeXtSmall
-from keras_cv.models.convnext import ConvNeXtTiny
-from keras_cv.models.convnext import ConvNeXtXLarge
-from keras_cv.models.darknet import DarkNet21
-from keras_cv.models.darknet import DarkNet53
-from keras_cv.models.densenet import DenseNet121
-from keras_cv.models.densenet import DenseNet169
-from keras_cv.models.densenet import DenseNet201
-from keras_cv.models.efficientnet_lite import EfficientNetLiteB0
-from keras_cv.models.efficientnet_lite import EfficientNetLiteB1
-from keras_cv.models.efficientnet_lite import EfficientNetLiteB2
-from keras_cv.models.efficientnet_lite import EfficientNetLiteB3
-from keras_cv.models.efficientnet_lite import EfficientNetLiteB4
-from keras_cv.models.efficientnet_v1 import EfficientNetB0
-from keras_cv.models.efficientnet_v1 import EfficientNetB1
-from keras_cv.models.efficientnet_v1 import EfficientNetB2
-from keras_cv.models.efficientnet_v1 import EfficientNetB3
-from keras_cv.models.efficientnet_v1 import EfficientNetB4
-from keras_cv.models.efficientnet_v1 import EfficientNetB5
-from keras_cv.models.efficientnet_v1 import EfficientNetB6
-from keras_cv.models.efficientnet_v1 import EfficientNetB7
-from keras_cv.models.mlp_mixer import MLPMixerB16
-from keras_cv.models.mlp_mixer import MLPMixerB32
-from keras_cv.models.mlp_mixer import MLPMixerL16
-from keras_cv.models.object_detection.faster_rcnn.faster_rcnn import FasterRCNN
 from keras_cv.models.object_detection.retinanet.retinanet import RetinaNet
 from keras_cv.models.object_detection.yolo_v8.yolo_v8_backbone import (
     YOLOV8Backbone,
@@ -142,42 +110,5 @@ from keras_cv.models.object_detection.yolo_v8.yolo_v8_detector import (
 from keras_cv.models.object_detection_3d.center_pillar import (
     MultiHeadCenterPillar,
 )
-from keras_cv.models.regnet import RegNetX002
-from keras_cv.models.regnet import RegNetX004
-from keras_cv.models.regnet import RegNetX006
-from keras_cv.models.regnet import RegNetX008
-from keras_cv.models.regnet import RegNetX016
-from keras_cv.models.regnet import RegNetX032
-from keras_cv.models.regnet import RegNetX040
-from keras_cv.models.regnet import RegNetX064
-from keras_cv.models.regnet import RegNetX080
-from keras_cv.models.regnet import RegNetX120
-from keras_cv.models.regnet import RegNetX160
-from keras_cv.models.regnet import RegNetX320
-from keras_cv.models.regnet import RegNetY002
-from keras_cv.models.regnet import RegNetY004
-from keras_cv.models.regnet import RegNetY006
-from keras_cv.models.regnet import RegNetY008
-from keras_cv.models.regnet import RegNetY016
-from keras_cv.models.regnet import RegNetY032
-from keras_cv.models.regnet import RegNetY040
-from keras_cv.models.regnet import RegNetY064
-from keras_cv.models.regnet import RegNetY080
-from keras_cv.models.regnet import RegNetY120
-from keras_cv.models.regnet import RegNetY160
-from keras_cv.models.regnet import RegNetY320
-from keras_cv.models.segmentation.deeplab import DeepLabV3
 from keras_cv.models.stable_diffusion import StableDiffusion
 from keras_cv.models.stable_diffusion import StableDiffusionV2
-from keras_cv.models.vgg16 import VGG16
-from keras_cv.models.vgg19 import VGG19
-from keras_cv.models.vit import ViTB16
-from keras_cv.models.vit import ViTB32
-from keras_cv.models.vit import ViTH16
-from keras_cv.models.vit import ViTH32
-from keras_cv.models.vit import ViTL16
-from keras_cv.models.vit import ViTL32
-from keras_cv.models.vit import ViTS16
-from keras_cv.models.vit import ViTS32
-from keras_cv.models.vit import ViTTiny16
-from keras_cv.models.vit import ViTTiny32
