+++ keras_cv/models/legacy/convmixer.py
@@ -23,8 +23,8 @@ import tensorflow as tf
 from tensorflow import keras
 from tensorflow.keras import layers
 
-from keras_cv.models import utils
-from keras_cv.models.weights import parse_weights
+from keras_cv.models.legacy import utils
+from keras_cv.models.legacy.weights import parse_weights
 
 MODEL_CONFIGS = {
     "ConvMixer_1536_20": {
