+++ keras_cv/models/legacy/convnext.py
@@ -24,7 +24,7 @@ from tensorflow.keras import backend
 from tensorflow.keras import layers
 
 from keras_cv.layers.regularization import StochasticDepth
-from keras_cv.models import utils
+from keras_cv.models.legacy import utils
 
 MODEL_CONFIGS = {
     "tiny": {
