+++ keras_cv/models/legacy/segmentation/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from keras_cv.models.segmentation.deeplab import DeepLabV3
+from keras_cv.models.legacy.segmentation.deeplab import DeepLabV3
