+++ keras_cv/models/legacy/vgg16.py
@@ -13,34 +13,73 @@
 # limitations under the License.
 # ==============================================================================
 
-"""VGG19 model for KerasCV.
+"""VGG16 model for KerasCV.
 Reference:
   - [Very Deep Convolutional Networks for Large-Scale Image Recognition](https://arxiv.org/abs/1409.1556)
     (ICLR 2015)
 """  # noqa: E501
+
 import tensorflow as tf
 from tensorflow import keras
 from tensorflow.keras import layers
 
-from keras_cv.models import utils
-from keras_cv.models.vgg16 import apply_vgg_block
+from keras_cv.models.legacy import utils
+
+
+def apply_vgg_block(
+    x,
+    num_layers,
+    filters,
+    kernel_size,
+    activation,
+    padding,
+    max_pool,
+    name,
+):
+    """
+    Applies VGG block
+    Args:
+        x: Tensor, input tensor to pass through network
+        num_layers: int, number of CNN layers in the block
+        filters: int, filter size of each CNN layer in block
+        kernel_size: int (or) tuple, kernel size for CNN layer in block
+        activation: str (or) callable, activation function for each CNN layer in
+            block
+        padding: str (or) callable, padding function for each CNN layer in block
+        max_pool: bool, whether to add MaxPooling2D layer at end of block
+        name: str, name of the block
+
+    Returns:
+        tf.Tensor
+    """
+    for num in range(1, num_layers + 1):
+        x = layers.Conv2D(
+            filters,
+            kernel_size,
+            activation=activation,
+            padding=padding,
+            name=f"{name}_conv{str(num)}",
+        )(x)
+    if max_pool:
+        x = layers.MaxPooling2D((2, 2), strides=(2, 2), name=f"{name}_pool")(x)
+    return x
 
 
 @keras.utils.register_keras_serializable(package="keras_cv.models")
-class VGG19(keras.Model):
+class VGG16(keras.Model):
     """
     Reference:
     - [Very Deep Convolutional Networks for Large-Scale Image Recognition](https://arxiv.org/abs/1409.1556)
         (ICLR 2015)
-    This class represents a Keras VGG19 model.
+    This class represents a Keras VGG16 model.
     Args:
       include_rescaling: bool, whether to rescale the inputs. If set to
         True, inputs will be passed through a `Rescaling(1/255.0)` layer.
       include_top: bool, whether to include the 3 fully-connected
         layers at the top of the network. If provided, num_classes must be
-        provided.
-      num_classes: int, optional number of classes to classify images into, only
-        to be specified if `include_top` is True.
+          provided.
+      num_classes: int, optional number of classes to classify images into,
+        only to be specified if `include_top` is True.
       weights: os.PathLike or None, one of `None` (random initialization), or a
         pretrained weight file path.
       input_shape: tuple, optional shape tuple, defaults to (224, 224, 3).
@@ -62,7 +101,7 @@ class VGG19(keras.Model):
         `classifier_activation=None` to return the logits of the "top" layer.
         When loading pretrained weights, `classifier_activation` can only
         be `None` or `"softmax"`.
-      name: (Optional) name to pass to the model, defaults to "VGG19".
+      name: (Optional) name to pass to the model, defaults to "VGG16".
     Returns:
       A `keras.Model` instance.
     """  # noqa: E501
@@ -77,7 +116,7 @@ class VGG19(keras.Model):
         input_shape=(224, 224, 3),
         pooling=None,
         classifier_activation="softmax",
-        name="VGG19",
+        name="VGG16",
         **kwargs,
     ):
         if weights and not tf.io.gfile.exists(weights):
@@ -129,7 +168,7 @@ class VGG19(keras.Model):
 
         x = apply_vgg_block(
             x=x,
-            num_layers=4,
+            num_layers=3,
             filters=256,
             kernel_size=(3, 3),
             activation="relu",
@@ -140,7 +179,7 @@ class VGG19(keras.Model):
 
         x = apply_vgg_block(
             x=x,
-            num_layers=4,
+            num_layers=3,
             filters=512,
             kernel_size=(3, 3),
             activation="relu",
@@ -151,7 +190,7 @@ class VGG19(keras.Model):
 
         x = apply_vgg_block(
             x=x,
-            num_layers=4,
+            num_layers=3,
             filters=512,
             kernel_size=(3, 3),
             activation="relu",
